getResult change in many class

+++ MiniSQLEngine/ClassDropDatabase.cs
@@ -1,6 +1,7 @@
 ﻿using MiniSQLEngine;
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -10,6 +11,7 @@ namespace MiniSQLEngine
     public class ClassDropDatabase : Query
     {
         private string name;
+        private string result;
 
         public ClassDropDatabase(string pName)
         {
@@ -20,11 +22,26 @@ namespace MiniSQLEngine
         {
             return "dropdb";
         }
+        public override string getResult()
+        {
+            return result;
+        }
 
         public override void Run(string dbname)
         {
-            string path = @"..//..//..//data//" + dbname;
-            System.IO.Directory.Delete(path, true);
+            Boolean hayerror = false;
+            if (Directory.Exists("..//..//..//data//" + dbname)==false)
+            {
+                result = Constants.Error;
+                hayerror = true;
+            }
+
+            if (hayerror == false)
+            {
+                string path = @"..//..//..//data//" + dbname;
+                System.IO.Directory.Delete(path, true);
+                result = Constants.DeleteDatabaseSuccess;
+            }
         }
         public string getDatabaseName()
         {
